// src/server/model/iServiceMaster.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IServiceMaster { 
    unit?: string;
    rate?: number;
    purchaseRtnAccountName?: string;
    gstUnit?: string;
    percentage?: boolean;
    sacCode?: string;
    serviceGroupName?: string;
    purchaseAccountId?: number;
    purchaseAccountName?: string;
    serviceGroupId?: number;
    purchaseRtnAccountId?: number;
    ledgerName?: string;
    ledgerId?: number;
    printName?: string;
    taxClassId?: number;
    taxClassName?: string;
    rtnLedgerName?: string;
    rtnLedgerId?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}